Derive LOCF (Last Observation Carried Forward) Records

Adds LOCF records as new observations for each 'by group' when the dataset does not contain observations for missed visits/time points.

derive_locf_records_(
  dataset,
  dataset_expected_obs,
  by_vars,
  order,
  keep_vars = NULL
)

Arguments

dataset

Input dataset

The columns specified by the by_vars and the order parameter are expected.

dataset_expected_obs

Expected observations dataset

Data frame with all the combinations of PARAMCD, PARAM, AVISIT, AVISITN, ... which are expected in the dataset is expected.

by_vars

Grouping variables

For each group defined by by_vars those observations from dataset_expected_obs are added to the output dataset which do not have a corresponding observation in the input dataset or for which AVAL is NA for the corresponding observation in the input dataset. Only variables specified in by_vars will be populated in the newly created records.

order

List of variables for sorting a dataset

The dataset is sorted by order before carrying the last observation forward (eg. AVAL) within each by_vars.

keep_vars

Variables that need carrying the last observation forward

Keep variables that need carrying the last observation forward other than AVAL (eg. PARAMN, VISITNUM). If by default NULL, only variables specified in by_vars and AVAL will be populated in the newly created records.

Value

The input dataset with the new "LOCF" observations added for each by_vars. Note, a variable will only be populated in the new parameter rows if it is specified in by_vars.

Details

For each group (with respect to the variables specified for the by_vars parameter) those observations from dataset_expected_obs are added to the output dataset

which do not have a corresponding observation in the input dataset or

for which AVAL is NA for the corresponding observation in the input dataset.

For the new observations, AVAL is set to the non-missing AVAL of the previous observation in the input dataset (when sorted by order) and DTYPE is set to "LOCF".

Author

G Gayatri